feat: Extract fhevm-sdk into its own package
This commit extracts the FHEVM related logic into a separate package, `@se-2/fhevm-sdk`. This includes the core FHEVM instance creation, decryption/encryption utilities, and React hooks. The existing Next.js application is updated to consume this new package.
