+++ packages/fhevm-sdk/package.json
@@ -1,0 +1,54 @@
+{
+  "name": "@se-2/fhevm-sdk",
+  "version": "0.1.0",
+  "private": true,
+  "type": "module",
+  "main": "dist/index.js",
+  "module": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "default": "./dist/index.js"
+    },
+    "./core": {
+      "types": "./dist/core/index.d.ts",
+      "default": "./dist/core/index.js"
+    },
+    "./storage": {
+      "types": "./dist/storage/index.d.ts",
+      "default": "./dist/storage/index.js"
+    },
+    "./types": {
+      "types": "./dist/fhevmTypes.d.ts",
+      "default": "./dist/fhevmTypes.js"
+    },
+    "./react": {
+      "types": "./dist/react/index.d.ts",
+      "default": "./dist/react/index.js"
+    }
+  },
+  "scripts": {
+    "build": "tsc -p tsconfig.json",
+    "watch": "tsc -p tsconfig.json --watch",
+    "clean": "rimraf dist"
+  },
+  "dependencies": {
+    "idb": "^8.0.3"
+  },
+  "peerDependencies": {
+    "@fhevm/mock-utils": "^0.0.1-6",
+    "@zama-fhe/relayer-sdk": "^0.1.2",
+    "ethers": "^6.13.4",
+    "react": "^18.0.0 || ^19.0.0"
+  },
+  "devDependencies": {
+    "@fhevm/mock-utils": "0.0.1-6",
+    "@types/node": "~18.19.50",
+    "@types/react": "~19.0.7",
+    "@zama-fhe/relayer-sdk": "0.1.2",
+    "ethers": "^6.13.7",
+    "react": "~19.0.0",
+    "typescript": "~5.8.2"
+  }
+}
