+++ packages/fhevm-sdk/src/fhevmTypes.ts
@@ -1,0 +1,39 @@
+import type { FhevmInstance as _FhevmInstance } from "@zama-fhe/relayer-sdk/bundle";
+import type { HandleContractPair as _HandleContractPair } from "@zama-fhe/relayer-sdk/bundle";
+import type { DecryptedResults as _DecryptedResults } from "@zama-fhe/relayer-sdk/bundle";
+import type { FhevmInstanceConfig as _FhevmInstanceConfig } from "@zama-fhe/relayer-sdk/web";
+
+export type FhevmInstance = _FhevmInstance;
+export type FhevmInstanceConfig = _FhevmInstanceConfig;
+export type HandleContractPair = _HandleContractPair;
+export type DecryptedResults = _DecryptedResults;
+
+export type FhevmDecryptionSignatureType = {
+  publicKey: string;
+  privateKey: string;
+  signature: string;
+  startTimestamp: number; // Unix timestamp in seconds
+  durationDays: number;
+  userAddress: `0x${string}`;
+  contractAddresses: `0x${string}`[];
+  eip712: EIP712Type;
+};
+
+export type EIP712Type = {
+  domain: {
+    chainId: number;
+    name: string;
+    verifyingContract: `0x${string}`;
+    version: string;
+  };
+
+  message: any;
+  primaryType: string;
+  types: {
+    [key: string]: {
+      name: string;
+      type: string;
+    }[];
+  };
+};
+
