+++ packages/fhevm-sdk/src/internal/PublicKeyStorage.ts
@@ -1,6 +1,9 @@
-export * from "@se-2/fhevm-sdk";
+import { DBSchema, IDBPDatabase, openDB } from "idb";
 
-// Compatibility shim: re-exported from SDK
+type FhevmStoredPublicKey = {
+  publicKeyId: string;
+  publicKey: Uint8Array;
+};
 
 type FhevmStoredPublicParams = {
   publicParamsId: string;
@@ -68,13 +71,13 @@ function assertFhevmStoredPublicKey(value: unknown): asserts value is FhevmStore
   if (!("publicKeyId" in value)) {
     throw new Error(`FhevmStoredPublicKey.publicKeyId does not exist`);
   }
-  if (typeof value.publicKeyId !== "string") {
+  if (typeof (value as any).publicKeyId !== "string") {
     throw new Error(`FhevmStoredPublicKey.publicKeyId must be a string`);
   }
   if (!("publicKey" in value)) {
     throw new Error(`FhevmStoredPublicKey.publicKey does not exist`);
   }
-  if (!(value.publicKey instanceof Uint8Array)) {
+  if (!((value as any).publicKey instanceof Uint8Array)) {
     throw new Error(`FhevmStoredPublicKey.publicKey must be a Uint8Array`);
   }
 }
@@ -89,15 +92,90 @@ function assertFhevmStoredPublicParams(value: unknown): asserts value is FhevmSt
   if (!("publicParamsId" in value)) {
     throw new Error(`FhevmStoredPublicParams.publicParamsId does not exist`);
   }
-  if (typeof value.publicParamsId !== "string") {
+  if (typeof (value as any).publicParamsId !== "string") {
     throw new Error(`FhevmStoredPublicParams.publicParamsId must be a string`);
   }
   if (!("publicParams" in value)) {
     throw new Error(`FhevmStoredPublicParams.publicParams does not exist`);
   }
-  if (!(value.publicParams instanceof Uint8Array)) {
+  if (!((value as any).publicParams instanceof Uint8Array)) {
     throw new Error(`FhevmStoredPublicParams.publicParams must be a Uint8Array`);
   }
 }
 
-// All logic moved to SDK
+export async function publicKeyStorageGet(aclAddress: `0x${string}`): Promise<{
+  publicKey?: FhevmInstanceConfigPublicKey;
+  publicParams: FhevmInstanceConfigPublicParams | null;
+}> {
+  const db = await _getDB();
+  if (!db) {
+    return { publicParams: null } as any;
+  }
+
+  let storedPublicKey: FhevmStoredPublicKey | null = null;
+  try {
+    const pk = await db.get("publicKeyStore", aclAddress);
+    if (pk?.value) {
+      assertFhevmStoredPublicKey(pk.value);
+      storedPublicKey = pk.value;
+    }
+  } catch {
+    // ignore
+  }
+
+  let storedPublicParams: FhevmStoredPublicParams | null = null;
+  try {
+    const pp = await db.get("paramsStore", aclAddress);
+    if (pp?.value) {
+      assertFhevmStoredPublicParams(pp.value);
+      storedPublicParams = pp.value;
+    }
+  } catch {
+    // ignore
+  }
+
+  const publicKeyData = storedPublicKey?.publicKey;
+  const publicKeyId = storedPublicKey?.publicKeyId;
+  const publicParams = storedPublicParams
+    ? {
+        "2048": storedPublicParams,
+      }
+    : null;
+
+  let publicKey: FhevmInstanceConfigPublicKey | undefined = undefined;
+
+  if (publicKeyId && publicKeyData) {
+    publicKey = {
+      id: publicKeyId,
+      data: publicKeyData,
+    };
+  }
+
+  return {
+    ...(publicKey !== undefined && { publicKey }),
+    publicParams,
+  } as any;
+}
+
+export async function publicKeyStorageSet(
+  aclAddress: `0x${string}`,
+  publicKey: FhevmStoredPublicKey | null,
+  publicParams: FhevmStoredPublicParams | null,
+) {
+  assertFhevmStoredPublicKey(publicKey);
+  assertFhevmStoredPublicParams(publicParams);
+
+  const db = await _getDB();
+  if (!db) {
+    return;
+  }
+
+  if (publicKey) {
+    await db.put("publicKeyStore", { acl: aclAddress, value: publicKey });
+  }
+
+  if (publicParams) {
+    await db.put("paramsStore", { acl: aclAddress, value: publicParams });
+  }
+}
+
