+++ packages/fhevm-sdk/src/internal/constants.ts
@@ -1,0 +1,2 @@
+export const SDK_CDN_URL = "https://cdn.zama.ai/relayer-sdk-js/0.1.2/relayer-sdk-js.umd.cjs";
+
