+++ packages/fhevm-sdk/src/internal/mock/fhevmMock.ts
@@ -1,0 +1,26 @@
+import { FhevmInstance } from "../../fhevmTypes";
+import { MockFhevmInstance } from "@fhevm/mock-utils";
+import { JsonRpcProvider } from "ethers";
+
+export const fhevmMockCreateInstance = async (parameters: {
+  rpcUrl: string;
+  chainId: number;
+  metadata: {
+    ACLAddress: `0x${string}`;
+    InputVerifierAddress: `0x${string}`;
+    KMSVerifierAddress: `0x${string}`;
+  };
+}): Promise<FhevmInstance> => {
+  const provider = new JsonRpcProvider(parameters.rpcUrl);
+  const instance = await MockFhevmInstance.create(provider, provider, {
+    aclContractAddress: parameters.metadata.ACLAddress,
+    chainId: parameters.chainId,
+    gatewayChainId: 55815,
+    inputVerifierContractAddress: parameters.metadata.InputVerifierAddress,
+    kmsContractAddress: parameters.metadata.KMSVerifierAddress,
+    verifyingContractAddressDecryption: "0x5ffdaAB0373E62E2ea2944776209aEf29E631A64",
+    verifyingContractAddressInputVerification: "0x812b06e1CDCE800494b79fFE4f925A504a9A9810",
+  });
+  return instance as unknown as FhevmInstance;
+};
+
