+++ packages/fhevm-sdk/src/react/useFHEEncryption.ts
@@ -1,7 +1,7 @@
 "use client";
 
 import { useCallback, useMemo } from "react";
-import { FhevmInstance } from "@se-2/fhevm-sdk";
+import { FhevmInstance } from "../fhevmTypes.js";
 import { RelayerEncryptedInput } from "@zama-fhe/relayer-sdk/web";
 import { ethers } from "ethers";
 
@@ -10,7 +10,6 @@ export type EncryptResult = {
   inputProof: Uint8Array;
 };
 
-// Map external encrypted integer type to RelayerEncryptedInput builder method
 export const getEncryptionMethod = (internalType: string) => {
   switch (internalType) {
     case "externalEbool":
@@ -30,21 +29,17 @@ export const getEncryptionMethod = (internalType: string) => {
     case "externalEaddress":
       return "addAddress" as const;
     default:
-      console.warn(`Unknown internalType: ${internalType}, defaulting to add64`);
       return "add64" as const;
   }
 };
 
-// Convert Uint8Array or hex-like string to 0x-prefixed hex string
 export const toHex = (value: Uint8Array | string): `0x${string}` => {
   if (typeof value === "string") {
     return (value.startsWith("0x") ? value : `0x${value}`) as `0x${string}`;
   }
-  // value is Uint8Array
   return ("0x" + Buffer.from(value).toString("hex")) as `0x${string}`;
 };
 
-// Build contract params from EncryptResult and ABI for a given function
 export const buildParamsFromAbi = (enc: EncryptResult, abi: any[], functionName: string): any[] => {
   const fn = abi.find((item: any) => item.type === "function" && item.name === functionName);
   if (!fn) throw new Error(`Function ABI not found for ${functionName}`);
@@ -59,11 +54,10 @@ export const buildParamsFromAbi = (enc: EncryptResult, abi: any[], functionName:
         return BigInt(raw as unknown as string);
       case "address":
       case "string":
-        return raw as unknown as string;
+        return (raw as unknown) as string;
       case "bool":
         return Boolean(raw);
       default:
-        console.warn(`Unknown ABI param type ${input.type}; passing as hex`);
         return toHex(raw);
     }
   });
@@ -76,20 +70,17 @@ export const useFHEEncryption = (params: {
 }) => {
   const { instance, ethersSigner, contractAddress } = params;
 
-  const canEncrypt = useMemo(
-    () => Boolean(instance && ethersSigner && contractAddress),
-    [instance, ethersSigner, contractAddress],
-  );
+  const canEncrypt = useMemo(() => Boolean(instance && ethersSigner && contractAddress), [instance, ethersSigner, contractAddress]);
 
   const encryptWith = useCallback(
     async (buildFn: (builder: RelayerEncryptedInput) => void): Promise<EncryptResult | undefined> => {
       if (!instance || !ethersSigner || !contractAddress) return undefined;
 
       const userAddress = await ethersSigner.getAddress();
-      const input = instance.createEncryptedInput(contractAddress, userAddress) as RelayerEncryptedInput;
+      const input = (instance as any).createEncryptedInput(contractAddress, userAddress) as RelayerEncryptedInput;
       buildFn(input);
       const enc = await input.encrypt();
-      return enc;
+      return enc as EncryptResult;
     },
     [instance, ethersSigner, contractAddress],
   );
@@ -99,3 +90,4 @@ export const useFHEEncryption = (params: {
     encryptWith,
   } as const;
 };
+
