+++ packages/fhevm-sdk/src/storage/index.ts
@@ -1,0 +1,2 @@
+export * from "./GenericStringStorage";
+
