+++ packages/nextjs/app/debug/_components/contract/DisplayVariable.tsx
@@ -7,8 +7,8 @@ import { Abi, AbiFunction } from "abitype";
 import { Address } from "viem";
 import { useReadContract } from "wagmi";
 import { ArrowPathIcon } from "@heroicons/react/24/outline";
-import { useFhevm } from "~~/fhevm/useFhevm";
-import { useFHEDecrypt } from "~~/hooks/fhevm/fhevm/useFHEDecrypt";
+import { useFhevm } from "@se-2/fhevm-sdk/react";
+import { useFHEDecrypt } from "@se-2/fhevm-sdk/react";
 import { useInMemoryStorage } from "~~/hooks/fhevm/fhevm/useInMemoryStorage";
 import { useWagmiEthers } from "~~/hooks/fhevm/useWagmiEthers";
 import { useAnimationConfig } from "~~/hooks/scaffold-eth";
