+++ packages/nextjs/app/fhevm/_components/FHECounterDemo.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { useMemo } from "react";
-import { useFhevm } from "../../../fhevm/useFhevm";
+import { useFhevm } from "@se-2/fhevm-sdk/react";
 import { useInMemoryStorage } from "../../../hooks/fhevm/fhevm/useInMemoryStorage";
 import { useFHECounterWagmi } from "../../../hooks/fhevm/useFHECounterWagmi";
 import { useAccount } from "wagmi";
