+++ packages/nextjs/fhevm/fhevmTypes.ts
@@ -1,35 +1,1 @@
-import type { FhevmInstance } from "@zama-fhe/relayer-sdk/bundle";
-import type { HandleContractPair } from "@zama-fhe/relayer-sdk/bundle";
-import type { DecryptedResults } from "@zama-fhe/relayer-sdk/bundle";
-import type { FhevmInstanceConfig } from "@zama-fhe/relayer-sdk/web";
-
-export type { FhevmInstance, FhevmInstanceConfig, HandleContractPair, DecryptedResults };
-
-export type FhevmDecryptionSignatureType = {
-  publicKey: string;
-  privateKey: string;
-  signature: string;
-  startTimestamp: number; // Unix timestamp in seconds
-  durationDays: number;
-  userAddress: `0x${string}`;
-  contractAddresses: `0x${string}`[];
-  eip712: EIP712Type;
-};
-
-export type EIP712Type = {
-  domain: {
-    chainId: number;
-    name: string;
-    verifyingContract: `0x${string}`;
-    version: string;
-  };
-
-  message: any;
-  primaryType: string;
-  types: {
-    [key: string]: {
-      name: string;
-      type: string;
-    }[];
-  };
-};
+export * from "@se-2/fhevm-sdk";
