+++ packages/nextjs/fhevm/internal/constants.ts
@@ -1,1 +1,1 @@
-export const SDK_CDN_URL = "https://cdn.zama.ai/relayer-sdk-js/0.1.2/relayer-sdk-js.umd.cjs";
+export * from "@se-2/fhevm-sdk";
