+++ packages/nextjs/fhevm/internal/mock/fhevmMock.ts
@@ -5,31 +5,4 @@
 // FHEVM MOCK LIB IN THE FINAL PRODUCTION BUNDLE!!
 //
 //////////////////////////////////////////////////////////////////////////
-import { FhevmInstance } from "../../fhevmTypes";
-import { MockFhevmInstance } from "@fhevm/mock-utils";
-import { JsonRpcProvider } from "ethers";
-
-export const fhevmMockCreateInstance = async (parameters: {
-  rpcUrl: string;
-  chainId: number;
-  metadata: {
-    ACLAddress: `0x${string}`;
-    InputVerifierAddress: `0x${string}`;
-    KMSVerifierAddress: `0x${string}`;
-  };
-}): Promise<FhevmInstance> => {
-  const provider = new JsonRpcProvider(parameters.rpcUrl);
-  const instance = await MockFhevmInstance.create(provider, provider, {
-    //aclContractAddress: "0x50157CFfD6bBFA2DECe204a89ec419c23ef5755D",
-    aclContractAddress: parameters.metadata.ACLAddress,
-    chainId: parameters.chainId,
-    gatewayChainId: 55815,
-    // inputVerifierContractAddress: "0x901F8942346f7AB3a01F6D7613119Bca447Bb030",
-    // kmsContractAddress: "0x1364cBBf2cDF5032C47d8226a6f6FBD2AFCDacAC",
-    inputVerifierContractAddress: parameters.metadata.InputVerifierAddress,
-    kmsContractAddress: parameters.metadata.KMSVerifierAddress,
-    verifyingContractAddressDecryption: "0x5ffdaAB0373E62E2ea2944776209aEf29E631A64",
-    verifyingContractAddressInputVerification: "0x812b06e1CDCE800494b79fFE4f925A504a9A9810",
-  });
-  return instance;
-};
+export * from "@se-2/fhevm-sdk";
