+++ packages/nextjs/package.json
@@ -18,6 +18,7 @@
   "dependencies": {
     "@heroicons/react": "~2.1.5",
     "@rainbow-me/rainbowkit": "2.2.8",
+    "@se-2/fhevm-sdk": "workspace:*",
     "@tanstack/react-query": "~5.59.15",
     "@uniswap/sdk-core": "~5.8.2",
     "@uniswap/v2-sdk": "~4.6.1",
